Removed env and moved collectors to the root for easier merge
